+++ backoff/backoff_test.go
@@ -1,4 +1,4 @@
-package util
+package backoff
 
 import (
 	"context"
@@ -85,7 +85,7 @@ func TestBackoff_NextDelay(t *testing.T) {
 		t.Run(testName, func(t *testing.T) {
 			t.Parallel()
 
-			b := NewBackoff(context.Background(), BackoffConfig{
+			b := NewBackoff(context.Background(), Config{
 				MinBackoff: testData.minBackoff,
 				MaxBackoff: testData.maxBackoff,
 				MaxRetries: len(testData.expectedRanges),
